+++ pkg/analysis_server/lib/src/lsp/handlers/handler_definition.dart
@@ -180,7 +180,7 @@ class DefinitionHandler extends MessageHandler<TextDocumentPositionParams,
   }
 
   /// Get the location of the code (excluding leading doc comments) for this element.
-  protocol.Location? _getCodeLocation(Element element) {
+  Future<protocol.Location?> _getCodeLocation(Element element) async {
     var codeElement = element;
     // For synthetic getters created for fields, we need to access the associated
     // variable to get the codeOffset/codeLength.
@@ -206,7 +206,7 @@ class DefinitionHandler extends MessageHandler<TextDocumentPositionParams,
     // Read the declaration so we can get the offset after the doc comments.
     // TODO(dantup): Skip this for parts (getParsedLibrary will throw), but find
     // a better solution.
-    final declaration = _parsedDeclaration(codeElement);
+    final declaration = await _parsedDeclaration(codeElement);
     var node = declaration?.node;
     if (node is VariableDeclaration) {
       node = node.parent;
@@ -244,9 +244,11 @@ class DefinitionHandler extends MessageHandler<TextDocumentPositionParams,
         : null;
   }
 
-  void _updateTargetsWithCodeLocations(NavigationCollectorImpl collector) {
+  Future<void> _updateTargetsWithCodeLocations(
+    NavigationCollectorImpl collector,
+  ) async {
     for (var targetToUpdate in collector.targetsToUpdate) {
-      var codeLocation = _getCodeLocation(targetToUpdate.element);
+      var codeLocation = await _getCodeLocation(targetToUpdate.element);
       if (codeLocation != null) {
         targetToUpdate.target
           ..codeOffset = codeLocation.offset
@@ -255,7 +257,9 @@ class DefinitionHandler extends MessageHandler<TextDocumentPositionParams,
     }
   }
 
-  static ElementDeclarationResult? _parsedDeclaration(Element element) {
+  static Future<ElementDeclarationResult?> _parsedDeclaration(
+    Element element,
+  ) async {
     var session = element.session;
     if (session == null) {
       return null;
@@ -266,7 +270,7 @@ class DefinitionHandler extends MessageHandler<TextDocumentPositionParams,
       return null;
     }
 
-    var parsedLibrary = session.getParsedLibrary(libraryPath);
+    var parsedLibrary = await session.getParsedLibrary2(libraryPath);
     if (parsedLibrary is! ParsedLibraryResult) {
       return null;
     }