+++ pkg/analyzer/test/src/dart/analysis/session_test.dart
@@ -44,6 +44,17 @@ class AnalysisSessionImpl_BazelWorkspaceTest
     expect(result.uri.toString(), 'package:dart.my/a.dart');
   }
 
+  void test_getParsedLibrary2_notFileOfUri() async {
+    var relPath = 'dart/my/lib/a.dart';
+    newFile('$workspaceRootPath/bazel-bin/$relPath');
+
+    var path = convertPath('$workspaceRootPath/$relPath');
+    var session = contextFor(path).currentSession;
+    var result = await session.getParsedLibrary2(path);
+    expect(result, isA<NotPathOfUriResult>());
+  }
+
+  @deprecated
   void test_getParsedLibrary_notFileOfUri() async {
     var relPath = 'dart/my/lib/a.dart';
     newFile('$workspaceRootPath/bazel-bin/$relPath');
@@ -150,13 +161,44 @@ class AnalysisSessionImplTest extends PubPackageResolutionTest {
     expect(errorsResult, isA<InvalidPathResult>());
   }
 
-  test_getFile_invalidPath_notAbsolute() async {
+  test_getFile2_inconsistent() async {
+    var test = newFile(testFilePath, content: '');
+    var session = contextFor(test.path).currentSession;
+    driverFor(test.path).changeFile(test.path);
+    expect(
+      () async => session.getFile2(test.path),
+      throwsA(isA<InconsistentAnalysisException>()),
+    );
+  }
+
+  test_getFile2_invalidPath_notAbsolute() async {
     var session = contextFor(testFilePath).currentSession;
-    var errorsResult = session.getFile('not_absolute.dart');
+    var errorsResult = await session.getFile2('not_absolute.dart');
     expect(errorsResult, isA<InvalidPathResult>());
   }
 
-  test_getFileSync_inconsistent() async {
+  test_getFile2_library() async {
+    var a = newFile('$testPackageLibPath/a.dart', content: '');
+
+    var session = contextFor(testFilePath).currentSession;
+    var file = await session.getFile2Valid(a.path);
+    expect(file.path, a.path);
+    expect(file.uri.toString(), 'package:test/a.dart');
+    expect(file.isPart, isFalse);
+  }
+
+  test_getFile2_part() async {
+    var a = newFile('$testPackageLibPath/a.dart', content: 'part of lib;');
+
+    var session = contextFor(testFilePath).currentSession;
+    var file = await session.getFile2Valid(a.path);
+    expect(file.path, a.path);
+    expect(file.uri.toString(), 'package:test/a.dart');
+    expect(file.isPart, isTrue);
+  }
+
+  @deprecated
+  test_getFile_inconsistent() async {
     var test = newFile(testFilePath, content: '');
     var session = contextFor(test.path).currentSession;
     driverFor(test.path).changeFile(test.path);
@@ -166,7 +208,15 @@ class AnalysisSessionImplTest extends PubPackageResolutionTest {
     );
   }
 
-  test_getFileSync_library() async {
+  @deprecated
+  test_getFile_invalidPath_notAbsolute() async {
+    var session = contextFor(testFilePath).currentSession;
+    var errorsResult = session.getFile('not_absolute.dart');
+    expect(errorsResult, isA<InvalidPathResult>());
+  }
+
+  @deprecated
+  test_getFile_library() async {
     var a = newFile('$testPackageLibPath/a.dart', content: '');
 
     var session = contextFor(testFilePath).currentSession;
@@ -176,7 +226,8 @@ class AnalysisSessionImplTest extends PubPackageResolutionTest {
     expect(file.isPart, isFalse);
   }
 
-  test_getFileSync_part() async {
+  @deprecated
+  test_getFile_part() async {
     var a = newFile('$testPackageLibPath/a.dart', content: 'part of lib;');
 
     var session = contextFor(testFilePath).currentSession;
@@ -216,6 +267,7 @@ class B {}
     expect(result, isA<CannotResolveUriResult>());
   }
 
+  @deprecated
   test_getParsedLibrary() async {
     var test = newFile('$testPackageLibPath/a.dart', content: r'''
 class A {}
@@ -236,6 +288,189 @@ class B {}
     }
   }
 
+  test_getParsedLibrary2() async {
+    var test = newFile('$testPackageLibPath/a.dart', content: r'''
+class A {}
+class B {}
+''');
+
+    var session = contextFor(testFilePath).currentSession;
+    var parsedLibrary = await session.getParsedLibrary2Valid(test.path);
+    expect(parsedLibrary.session, session);
+
+    expect(parsedLibrary.units, hasLength(1));
+    {
+      var parsedUnit = parsedLibrary.units[0];
+      expect(parsedUnit.session, session);
+      expect(parsedUnit.path, test.path);
+      expect(parsedUnit.uri, Uri.parse('package:test/a.dart'));
+      expect(parsedUnit.unit.declarations, hasLength(2));
+    }
+  }
+
+  test_getParsedLibrary2_getElementDeclaration_class() async {
+    var test = newFile(testFilePath, content: r'''
+class A {}
+class B {}
+''');
+
+    var session = contextFor(testFilePath).currentSession;
+    var libraryResult = await session.getLibraryByUriValid(
+      'package:test/test.dart',
+    );
+    var parsedLibrary = await session.getParsedLibrary2Valid(test.path);
+
+    var element = libraryResult.element.getType('A')!;
+    var declaration = parsedLibrary.getElementDeclaration(element)!;
+    var node = declaration.node as ClassDeclaration;
+    expect(node.name.name, 'A');
+    expect(node.offset, 0);
+    expect(node.length, 10);
+  }
+
+  test_getParsedLibrary2_getElementDeclaration_notThisLibrary() async {
+    var test = newFile(testFilePath, content: '');
+
+    var session = contextFor(testFilePath).currentSession;
+    var resolvedUnit =
+        await session.getResolvedUnit(test.path) as ResolvedUnitResult;
+    var typeProvider = resolvedUnit.typeProvider;
+    var intClass = typeProvider.intType.element;
+
+    var parsedLibrary = await session.getParsedLibrary2Valid(test.path);
+
+    expect(() {
+      parsedLibrary.getElementDeclaration(intClass);
+    }, throwsArgumentError);
+  }
+
+  test_getParsedLibrary2_getElementDeclaration_synthetic() async {
+    var test = newFile(testFilePath, content: r'''
+int foo = 0;
+''');
+
+    var session = contextFor(testFilePath).currentSession;
+    var parsedLibrary = await session.getParsedLibrary2Valid(test.path);
+
+    var unitResult = await session.getUnitElementValid(test.path);
+    var fooElement = unitResult.element.topLevelVariables[0];
+    expect(fooElement.name, 'foo');
+
+    // We can get the variable element declaration.
+    var fooDeclaration = parsedLibrary.getElementDeclaration(fooElement)!;
+    var fooNode = fooDeclaration.node as VariableDeclaration;
+    expect(fooNode.name.name, 'foo');
+    expect(fooNode.offset, 4);
+    expect(fooNode.length, 7);
+    expect(fooNode.name.staticElement, isNull);
+
+    // Synthetic elements don't have nodes.
+    expect(parsedLibrary.getElementDeclaration(fooElement.getter!), isNull);
+    expect(parsedLibrary.getElementDeclaration(fooElement.setter!), isNull);
+  }
+
+  test_getParsedLibrary2_inconsistent() async {
+    var test = newFile(testFilePath, content: '');
+    var session = contextFor(test.path).currentSession;
+    driverFor(test.path).changeFile(test.path);
+    expect(
+      () => session.getParsedLibrary2(test.path),
+      throwsA(isA<InconsistentAnalysisException>()),
+    );
+  }
+
+  test_getParsedLibrary2_invalidPartUri() async {
+    var test = newFile(testFilePath, content: r'''
+part 'a.dart';
+part ':[invalid uri].dart';
+part 'c.dart';
+''');
+
+    var session = contextFor(testFilePath).currentSession;
+    var parsedLibrary = await session.getParsedLibrary2Valid(test.path);
+
+    expect(parsedLibrary.units, hasLength(3));
+    expect(
+      parsedLibrary.units[0].path,
+      convertPath('/home/test/lib/test.dart'),
+    );
+    expect(
+      parsedLibrary.units[1].path,
+      convertPath('/home/test/lib/a.dart'),
+    );
+    expect(
+      parsedLibrary.units[2].path,
+      convertPath('/home/test/lib/c.dart'),
+    );
+  }
+
+  test_getParsedLibrary2_invalidPath_notAbsolute() async {
+    var session = contextFor(testFilePath).currentSession;
+    var result = await session.getParsedLibrary2('not_absolute.dart');
+    expect(result, isA<InvalidPathResult>());
+  }
+
+  test_getParsedLibrary2_notLibrary() async {
+    var test = newFile(testFilePath, content: 'part of "a.dart";');
+    var session = contextFor(testFilePath).currentSession;
+    var result = await session.getParsedLibrary2(test.path);
+    expect(result, isA<NotLibraryButPartResult>());
+  }
+
+  test_getParsedLibrary2_parts() async {
+    var aContent = r'''
+part 'b.dart';
+part 'c.dart';
+
+class A {}
+''';
+
+    var bContent = r'''
+part of 'a.dart';
+
+class B1 {}
+class B2 {}
+''';
+
+    var cContent = r'''
+part of 'a.dart';
+
+class C1 {}
+class C2 {}
+class C3 {}
+''';
+
+    var a = newFile('$testPackageLibPath/a.dart', content: aContent);
+    var b = newFile('$testPackageLibPath/b.dart', content: bContent);
+    var c = newFile('$testPackageLibPath/c.dart', content: cContent);
+
+    var session = contextFor(testFilePath).currentSession;
+    var parsedLibrary = await session.getParsedLibrary2Valid(a.path);
+    expect(parsedLibrary.units, hasLength(3));
+
+    {
+      var aUnit = parsedLibrary.units[0];
+      expect(aUnit.path, a.path);
+      expect(aUnit.uri, Uri.parse('package:test/a.dart'));
+      expect(aUnit.unit.declarations, hasLength(1));
+    }
+
+    {
+      var bUnit = parsedLibrary.units[1];
+      expect(bUnit.path, b.path);
+      expect(bUnit.uri, Uri.parse('package:test/b.dart'));
+      expect(bUnit.unit.declarations, hasLength(2));
+    }
+
+    {
+      var cUnit = parsedLibrary.units[2];
+      expect(cUnit.path, c.path);
+      expect(cUnit.uri, Uri.parse('package:test/c.dart'));
+      expect(cUnit.unit.declarations, hasLength(3));
+    }
+  }
+
+  @deprecated
   test_getParsedLibrary_getElementDeclaration_class() async {
     var test = newFile(testFilePath, content: r'''
 class A {}
@@ -256,6 +491,7 @@ class B {}
     expect(node.length, 10);
   }
 
+  @deprecated
   test_getParsedLibrary_getElementDeclaration_notThisLibrary() async {
     var test = newFile(testFilePath, content: '');
 
@@ -272,6 +508,7 @@ class B {}
     }, throwsArgumentError);
   }
 
+  @deprecated
   test_getParsedLibrary_getElementDeclaration_synthetic() async {
     var test = newFile(testFilePath, content: r'''
 int foo = 0;
@@ -297,6 +534,7 @@ int foo = 0;
     expect(parsedLibrary.getElementDeclaration(fooElement.setter!), isNull);
   }
 
+  @deprecated
   test_getParsedLibrary_inconsistent() async {
     var test = newFile(testFilePath, content: '');
     var session = contextFor(test.path).currentSession;
@@ -307,6 +545,7 @@ int foo = 0;
     );
   }
 
+  @deprecated
   test_getParsedLibrary_invalidPartUri() async {
     var test = newFile(testFilePath, content: r'''
 part 'a.dart';
@@ -332,18 +571,21 @@ part 'c.dart';
     );
   }
 
+  @deprecated
   test_getParsedLibrary_invalidPath_notAbsolute() async {
     var session = contextFor(testFilePath).currentSession;
     var result = session.getParsedLibrary('not_absolute.dart');
     expect(result, isA<InvalidPathResult>());
   }
 
+  @deprecated
   test_getParsedLibrary_notLibrary() async {
     var test = newFile(testFilePath, content: 'part of "a.dart";');
     var session = contextFor(testFilePath).currentSession;
     expect(session.getParsedLibrary(test.path), isA<NotLibraryButPartResult>());
   }
 
+  @deprecated
   test_getParsedLibrary_parts() async {
     var aContent = r'''
 part 'b.dart';
@@ -397,6 +639,7 @@ class C3 {}
     }
   }
 
+  @deprecated
   test_getParsedLibraryByElement() async {
     var test = newFile(testFilePath, content: '');
 
@@ -418,6 +661,43 @@ class C3 {}
     }
   }
 
+  test_getParsedLibraryByElement2() async {
+    var test = newFile(testFilePath, content: '');
+
+    var session = contextFor(testFilePath).currentSession;
+    var libraryResult = await session.getLibraryByUriValid(
+      'package:test/test.dart',
+    );
+    var element = libraryResult.element;
+
+    var parsedLibrary = await session.getParsedLibraryByElement2Valid(element);
+    expect(parsedLibrary.session, session);
+    expect(parsedLibrary.units, hasLength(1));
+
+    {
+      var unit = parsedLibrary.units[0];
+      expect(unit.path, test.path);
+      expect(unit.uri, Uri.parse('package:test/test.dart'));
+      expect(unit.unit, isNotNull);
+    }
+  }
+
+  test_getParsedLibraryByElement2_differentSession() async {
+    newFile(testFilePath, content: '');
+
+    var session = contextFor(testFilePath).currentSession;
+    var libraryResult = await session.getLibraryByUriValid(
+      'package:test/test.dart',
+    );
+    var element = libraryResult.element;
+
+    var aaaSession = contextFor('$workspaceRootPath/aaa').currentSession;
+
+    var result = await aaaSession.getParsedLibraryByElement2(element);
+    expect(result, isA<NotElementOfThisSessionResult>());
+  }
+
+  @deprecated
   test_getParsedLibraryByElement_differentSession() async {
     newFile(testFilePath, content: '');
 
@@ -433,6 +713,7 @@ class C3 {}
     expect(result, isA<NotElementOfThisSessionResult>());
   }
 
+  @deprecated
   test_getParsedUnit() async {
     var test = newFile(testFilePath, content: r'''
 class A {}
@@ -447,6 +728,37 @@ class B {}
     expect(unitResult.unit.declarations, hasLength(2));
   }
 
+  test_getParsedUnit2() async {
+    var test = newFile(testFilePath, content: r'''
+class A {}
+class B {}
+''');
+
+    var session = contextFor(testFilePath).currentSession;
+    var unitResult = await session.getParsedUnit2Valid(test.path);
+    expect(unitResult.session, session);
+    expect(unitResult.path, test.path);
+    expect(unitResult.uri, Uri.parse('package:test/test.dart'));
+    expect(unitResult.unit.declarations, hasLength(2));
+  }
+
+  test_getParsedUnit2_inconsistent() async {
+    var test = newFile(testFilePath, content: '');
+    var session = contextFor(test.path).currentSession;
+    driverFor(test.path).changeFile(test.path);
+    expect(
+      () => session.getParsedUnit2(test.path),
+      throwsA(isA<InconsistentAnalysisException>()),
+    );
+  }
+
+  test_getParsedUnit2_invalidPath_notAbsolute() async {
+    var session = contextFor(testFilePath).currentSession;
+    var result = await session.getParsedUnit2('not_absolute.dart');
+    expect(result, isA<InvalidPathResult>());
+  }
+
+  @deprecated
   test_getParsedUnit_inconsistent() async {
     var test = newFile(testFilePath, content: '');
     var session = contextFor(test.path).currentSession;
@@ -457,6 +769,7 @@ class B {}
     );
   }
 
+  @deprecated
   test_getParsedUnit_invalidPath_notAbsolute() async {
     var session = contextFor(testFilePath).currentSession;
     var result = session.getParsedUnit('not_absolute.dart');
@@ -705,6 +1018,11 @@ extension on AnalysisSession {
     return await getErrors(path) as ErrorsResult;
   }
 
+  Future<FileResult> getFile2Valid(String path) async {
+    return await getFile2(path) as FileResult;
+  }
+
+  @deprecated
   FileResult getFileValid(String path) {
     return getFile(path) as FileResult;
   }
@@ -713,14 +1031,31 @@ extension on AnalysisSession {
     return await getLibraryByUri(path) as LibraryElementResult;
   }
 
+  Future<ParsedLibraryResult> getParsedLibrary2Valid(String path) async {
+    return await getParsedLibrary2(path) as ParsedLibraryResult;
+  }
+
+  Future<ParsedLibraryResult> getParsedLibraryByElement2Valid(
+    LibraryElement element,
+  ) async {
+    return await getParsedLibraryByElement2(element) as ParsedLibraryResult;
+  }
+
+  @deprecated
   ParsedLibraryResult getParsedLibraryByElementValid(LibraryElement element) {
     return getParsedLibraryByElement(element) as ParsedLibraryResult;
   }
 
+  @deprecated
   ParsedLibraryResult getParsedLibraryValid(String path) {
     return getParsedLibrary(path) as ParsedLibraryResult;
   }
 
+  Future<ParsedUnitResult> getParsedUnit2Valid(String path) async {
+    return await getParsedUnit2(path) as ParsedUnitResult;
+  }
+
+  @deprecated
   ParsedUnitResult getParsedUnitValid(String path) {
     return getParsedUnit(path) as ParsedUnitResult;
   }